Update main.py
ship shooter 2.0

+++ main.py
@@ -1,4 +1,5 @@
 import pygame
+import time
 pygame.init()
 screen=pygame.display.set_mode((800,600))
 player1=pygame.image.load("C:\Aahana\Game Dev 2\lesson 6\Things\player 1.png")
@@ -8,6 +9,8 @@ player1x=50
 player2x=700
 player1y=300
 player2y=300
+yellow_health=10
+red_health=10
 bg=pygame.transform.scale(bg,(800,600))
 player1=pygame.transform.scale(player1,(50,50))
 player2=pygame.transform.scale(player2,(50,50))
@@ -16,13 +19,19 @@ player1=pygame.transform.rotate(player1,90)
 player2=pygame.transform.rotate(player2,270)
 yellowb=[]
 redb=[]
-RED_HEALTH=5
-YELLOW_HEALTH=6
-red=pygame.Rect(player1x,player1y,10,5)
-yellow=pygame.Rect(player2x,player2y,10,5)
+red=pygame.Rect(player1x,player1y,50,50)
+yellow=pygame.Rect(player2x,player2y,50,50)
+font=pygame.font.SysFont("brannboll script",40)
+
+    
 while True:
+    text1=font.render(str(red_health),True,"red")
+    text2=font.render(str(yellow_health),True,"orange")
     screen.blit(bg,(0,0))
     screen.blit(player1,(player1x,player1y))
+    screen.blit(text1,(15,15))
+    screen.blit(text2,(720,15))
+    screen.blit(player2,(player2x,player2y))
     screen.blit(player2,(player2x,player2y))
     pygame.draw.rect(screen,"black",border)
     for b in yellowb:
@@ -30,12 +39,27 @@ while True:
         b.x-=10
         if red.colliderect (b):
             yellowb.remove(b)
+            red_health=red_health-1
+    if yellow_health <0:
+        text4=font.render("red wins!!!!!",True,"purple")
+        screen.blit(text4,(400,300))
+        pygame.display.update()
+        time.sleep(3)
+        break
+            
 
     for r in redb:
         pygame.draw.rect(screen,"red",r)
         r.x+=5
         if yellow.colliderect(r):
             redb.remove(r)
+            yellow_health=yellow_health-1
+    if red_health <0:
+        text3=font.render("yellow wins!",True,"purple")
+        screen.blit(text3,(400,300)) 
+        pygame.display.update()
+        time.sleep(3) 
+        break
     pygame.display.update()
     for event in pygame.event.get():
         if event.type==pygame.KEYDOWN:
@@ -60,7 +84,8 @@ while True:
                 player2x=player2x-15
             if event.key==pygame.K_d:
                 player2y=player2y+15
-            if event.key==pygame.K_LSUPER:
+            if event.key==pygame.K_LALT:
                 bullet=pygame.Rect(player2x,player2y+25,10,5)
                 yellowb.append(bullet)
-                #print health
+                #Practice: 
+#Make the bumble bee game done in pgzero in pygame
